Reorganize folder and file structure to follow MVC principles.
