+++ src/main/java/com/brewandbite/model/inventory/Ingredient.java
@@ -1,10 +1,10 @@
-package com.brewandbite.model;
+package com.brewandbite.model.inventory;
 
 public class Ingredient {
+
     private String name;
     private int quantity;
     private String unit; // "grams", "ml", "pieces"
-
 
     public String getName() {
         return name;
@@ -30,17 +30,16 @@ public class Ingredient {
         this.unit = unit;
     }
 
-
-    public void reduceQuantity(int amount){
+    public void reduceQuantity(int amount) {
         this.quantity = this.quantity - amount;
     }
 
     @Override
     public String toString() {
-        return "Ingredient{" +
-                "name='" + name + '\'' +
-                ", quantity=" + quantity +
-                ", unit='" + unit + '\'' +
-                '}';
+        return "Ingredient{"
+                + "name='" + name + '\''
+                + ", quantity=" + quantity
+                + ", unit='" + unit + '\''
+                + '}';
     }
 }
