+++ src/main/java/com/brewandbite/model/items/Beverage.java
@@ -1,10 +1,8 @@
-package com.brewandbite.max;
+package com.brewandbite.model.items;
 
 import java.util.ArrayList;
 import java.util.Collections;
 import java.util.List;
-
-import com.brewandbite.model.MenuItem;
 
 public class Beverage extends MenuItem {
